Ensure the literal types for _DoubleWide are able to be at least 64 bits on 32 bit platforms

+++ stdlib/public/core/DoubleWide.swift
@@ -636,7 +636,7 @@ extension _DoubleWidth: FixedWidthInteger {
   }
   
   @inlinable
-  init(integerLiteral x: Int) {
+  init(integerLiteral x: Int64) {
     self.init(x)
   }
